// src/api/generated/model/todoDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { TodoDtoCat } from './todoDtoCat';
import type { TodoDtoStatus } from './todoDtoStatus';
import type { TodoDtoPrio } from './todoDtoPrio';

export interface TodoDto {
  id?: string;
  title?: string;
  cat?: TodoDtoCat;
  status?: TodoDtoStatus;
  prio?: TodoDtoPrio;
}