// src/api/generated/model/todoDtoPrio.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type TodoDtoPrio = typeof TodoDtoPrio[keyof typeof TodoDtoPrio];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TodoDtoPrio = {
  hoch: 'hoch',
  mittel: 'mittel',
  niedrig: 'niedrig',
} as const;