// src/api/generated/model/todoDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type TodoDtoStatus = typeof TodoDtoStatus[keyof typeof TodoDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TodoDtoStatus = {
  todo: 'todo',
  doing: 'doing',
  done: 'done',
} as const;